:recycle: Split ConsumerApiService into UseCase oriented services
