--- src/consumer_api_service.ts
@@ -1,963 +1,0 @@
-import * as EssentialProjectErrors from '@essential-projects/errors_ts';
-import {IEventAggregator, Subscription} from '@essential-projects/event_aggregator_contracts';
-import {IIAMService, IIdentity} from '@essential-projects/iam_contracts';
-import {DataModels, IConsumerApi, Messages} from '@process-engine/consumer_api_contracts';
-import {
-  BpmnType,
-  FlowNodeInstance,
-  FlowNodeInstanceState,
-  IFlowNodeInstanceService,
-  ProcessToken,
-  ProcessTokenType,
-} from '@process-engine/flow_node_instance.contracts';
-import {
-  IProcessModelFacadeFactory,
-  ActivityFinishedMessage as InternalActivityFinishedMessage,
-  FinishEmptyActivityMessage as InternalFinishEmptyActivityMessage,
-  FinishManualTaskMessage as InternalFinishManualTaskMessage,
-  FinishUserTaskMessage as InternalFinishUserTaskMessage,
-  UserTaskFinishedMessage as InternalUserTaskFinishedMessage,
-} from '@process-engine/process_engine_contracts';
-import {IProcessModelUseCases, Model} from '@process-engine/process_model.contracts';
-
-import {IProcessModelExecutionAdapter, NotificationAdapter} from './adapters/index';
-import {
-  EmptyActivityConverter,
-  EventConverter,
-  ManualTaskConverter,
-  ProcessInstanceConverter,
-  ProcessModelConverter,
-  UserTaskConverter,
-} from './converters/index';
-
-export class ConsumerApiService implements IConsumerApi {
-
-  private readonly eventAggregator: IEventAggregator;
-  private readonly eventConverter: EventConverter;
-  private readonly flowNodeInstanceService: IFlowNodeInstanceService;
-  private readonly iamService: IIAMService;
-  private readonly processModelExecutionAdapter: IProcessModelExecutionAdapter;
-  private readonly processModelFacadeFactory: IProcessModelFacadeFactory;
-  private readonly processModelUseCase: IProcessModelUseCases;
-
-  private readonly notificationAdapter: NotificationAdapter;
-
-  private readonly emptyActivityConverter: EmptyActivityConverter;
-  private readonly userTaskConverter: UserTaskConverter;
-  private readonly manualTaskConverter: ManualTaskConverter;
-  private readonly processInstanceConverter: ProcessInstanceConverter;
-  private readonly processModelConverter: ProcessModelConverter;
-
-  private readonly canTriggerMessagesClaim = 'can_trigger_messages';
-  private readonly canTriggerSignalsClaim = 'can_trigger_signals';
-  private readonly canSubscribeToEventsClaim = 'can_subscribe_to_events';
-
-  constructor(
-    eventAggregator: IEventAggregator,
-    flowNodeInstanceService: IFlowNodeInstanceService,
-    iamService: IIAMService,
-    processModelExecutionAdapter: IProcessModelExecutionAdapter,
-    processModelFacadeFactory: IProcessModelFacadeFactory,
-    processModelUseCase: IProcessModelUseCases,
-    notificationAdapter: NotificationAdapter,
-    emptyActivityConverter: EmptyActivityConverter,
-    eventConverter: EventConverter,
-    userTaskConverter: UserTaskConverter,
-    manualTaskConverter: ManualTaskConverter,
-    processInstanceConverter: ProcessInstanceConverter,
-    processModelConverter: ProcessModelConverter,
-  ) {
-    this.eventAggregator = eventAggregator;
-    this.flowNodeInstanceService = flowNodeInstanceService;
-    this.iamService = iamService;
-    this.processModelExecutionAdapter = processModelExecutionAdapter;
-    this.processModelFacadeFactory = processModelFacadeFactory;
-    this.processModelUseCase = processModelUseCase;
-
-    this.notificationAdapter = notificationAdapter;
-
-    this.emptyActivityConverter = emptyActivityConverter;
-    this.eventConverter = eventConverter;
-    this.userTaskConverter = userTaskConverter;
-    this.manualTaskConverter = manualTaskConverter;
-    this.processInstanceConverter = processInstanceConverter;
-    this.processModelConverter = processModelConverter;
-  }
-
-  // Notifications
-  public async onActivityReached(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnActivityReachedCallback,
-    subscribeOnce: boolean = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onActivityReached(identity, callback, subscribeOnce);
-  }
-
-  public async onActivityFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnActivityFinishedCallback,
-    subscribeOnce: boolean = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onActivityFinished(identity, callback, subscribeOnce);
-  }
-
-  // ------------ For backwards compatibility only
-
-  public async onCallActivityWaiting(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnCallActivityWaitingCallback,
-    subscribeOnce: boolean = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onCallActivityWaiting(identity, callback, subscribeOnce);
-  }
-
-  public async onCallActivityFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnCallActivityFinishedCallback,
-    subscribeOnce: boolean = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onCallActivityFinished(identity, callback, subscribeOnce);
-  }
-
-  // ------------
-
-  public async onEmptyActivityWaiting(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnEmptyActivityWaitingCallback,
-    subscribeOnce?: boolean,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onEmptyActivityWaiting(identity, callback, subscribeOnce);
-  }
-
-  public async onEmptyActivityFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnEmptyActivityFinishedCallback,
-    subscribeOnce?: boolean,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onEmptyActivityFinished(identity, callback, subscribeOnce);
-  }
-
-  public async onEmptyActivityForIdentityWaiting(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnEmptyActivityWaitingCallback,
-    subscribeOnce?: boolean,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onEmptyActivityForIdentityWaiting(identity, callback, subscribeOnce);
-  }
-
-  public async onEmptyActivityForIdentityFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnEmptyActivityFinishedCallback,
-    subscribeOnce?: boolean,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onEmptyActivityForIdentityFinished(identity, callback, subscribeOnce);
-  }
-
-  public async onUserTaskWaiting(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnUserTaskWaitingCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onUserTaskWaiting(identity, callback, subscribeOnce);
-  }
-
-  public async onUserTaskFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnUserTaskFinishedCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onUserTaskFinished(identity, callback, subscribeOnce);
-  }
-
-  public async onUserTaskForIdentityWaiting(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnUserTaskWaitingCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onUserTaskForIdentityWaiting(identity, callback, subscribeOnce);
-  }
-
-  public async onUserTaskForIdentityFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnUserTaskFinishedCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onUserTaskForIdentityFinished(identity, callback, subscribeOnce);
-  }
-
-  public async onBoundaryEventTriggered(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnBoundaryEventTriggeredCallback,
-    subscribeOnce: boolean = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onBoundaryEventTriggered(identity, callback, subscribeOnce);
-  }
-
-  public async onIntermediateCatchEventReached(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnIntermediateCatchEventReachedCallback,
-    subscribeOnce: boolean = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onIntermediateCatchEventReached(identity, callback, subscribeOnce);
-  }
-
-  public async onIntermediateCatchEventFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnIntermediateCatchEventFinishedCallback,
-    subscribeOnce: boolean = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onIntermediateCatchEventFinished(identity, callback, subscribeOnce);
-  }
-
-  public async onIntermediateThrowEventTriggered(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnIntermediateThrowEventTriggeredCallback,
-    subscribeOnce: boolean = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onIntermediateThrowEventTriggered(identity, callback, subscribeOnce);
-  }
-
-  public async onManualTaskWaiting(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnManualTaskWaitingCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onManualTaskWaiting(identity, callback, subscribeOnce);
-  }
-
-  public async onManualTaskFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnManualTaskFinishedCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onManualTaskFinished(identity, callback, subscribeOnce);
-  }
-
-  public async onManualTaskForIdentityWaiting(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnManualTaskWaitingCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onManualTaskForIdentityWaiting(identity, callback, subscribeOnce);
-  }
-
-  public async onManualTaskForIdentityFinished(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnManualTaskFinishedCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onManualTaskForIdentityFinished(identity, callback, subscribeOnce);
-  }
-
-  public async onProcessStarted(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnProcessStartedCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onProcessStarted(identity, callback, subscribeOnce);
-  }
-
-  public async onProcessWithProcessModelIdStarted(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnProcessStartedCallback,
-    processModelId: string,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onProcessWithProcessModelIdStarted(identity, callback, processModelId, subscribeOnce);
-  }
-
-  public async onProcessEnded(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnProcessEndedCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onProcessEnded(identity, callback, subscribeOnce);
-  }
-
-  public async onProcessTerminated(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnProcessTerminatedCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onProcessTerminated(identity, callback, subscribeOnce);
-  }
-
-  public async onProcessError(
-    identity: IIdentity,
-    callback: Messages.CallbackTypes.OnProcessErrorCallback,
-    subscribeOnce = false,
-  ): Promise<Subscription> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    return this.notificationAdapter.onProcessError(identity, callback, subscribeOnce);
-  }
-
-  public async removeSubscription(identity: IIdentity, subscription: Subscription): Promise<void> {
-    await this.iamService.ensureHasClaim(identity, this.canSubscribeToEventsClaim);
-
-    this.notificationAdapter.removeSubscription(subscription);
-  }
-
-  // Process models and instances
-  public async getProcessModels(identity: IIdentity): Promise<DataModels.ProcessModels.ProcessModelList> {
-
-    const processModels = await this.processModelUseCase.getProcessModels(identity);
-    const consumerApiProcessModels = processModels.map((processModel: Model.Process): DataModels.ProcessModels.ProcessModel => {
-      return this.processModelConverter.convertProcessModel(processModel);
-    });
-
-    return {
-      processModels: consumerApiProcessModels,
-    };
-  }
-
-  public async getProcessModelById(identity: IIdentity, processModelId: string): Promise<DataModels.ProcessModels.ProcessModel> {
-
-    const processModel = await this.processModelUseCase.getProcessModelById(identity, processModelId);
-    const consumerApiProcessModel = this.processModelConverter.convertProcessModel(processModel);
-
-    return consumerApiProcessModel;
-  }
-
-  public async getProcessModelByProcessInstanceId(identity: IIdentity, processInstanceId: string): Promise<DataModels.ProcessModels.ProcessModel> {
-
-    const processModel = await this.processModelUseCase.getProcessModelByProcessInstanceId(identity, processInstanceId);
-    const consumerApiProcessModel = this.processModelConverter.convertProcessModel(processModel);
-
-    return consumerApiProcessModel;
-  }
-
-  public async startProcessInstance(
-    identity: IIdentity,
-    processModelId: string,
-    payload: DataModels.ProcessModels.ProcessStartRequestPayload,
-    startCallbackType?: DataModels.ProcessModels.StartCallbackType,
-    startEventId?: string,
-    endEventId?: string,
-  ): Promise<DataModels.ProcessModels.ProcessStartResponsePayload> {
-
-    return this
-      .processModelExecutionAdapter
-      .startProcessInstance(identity, processModelId, payload, startCallbackType, startEventId, endEventId);
-  }
-
-  public async getProcessResultForCorrelation(
-    identity: IIdentity,
-    correlationId: string,
-    processModelId: string,
-  ): Promise<Array<DataModels.CorrelationResult>> {
-
-    const processModel =
-      await this.processModelUseCase.getProcessModelById(identity, processModelId);
-
-    // First retreive all EndEvents the user can access.
-    const processModelFacade = this.processModelFacadeFactory.create(processModel);
-    const userAccessibleEndEvents = processModelFacade.getEndEvents();
-
-    // Get all FlowNodeInstances that were run in the Correlation.
-    const flowNodeInstances = await this.flowNodeInstanceService.queryByCorrelation(correlationId);
-
-    const noResultsFound = !flowNodeInstances || flowNodeInstances.length === 0;
-    if (noResultsFound) {
-      throw new EssentialProjectErrors.NotFoundError(`No process results for correlation with id '${correlationId}' found.`);
-    }
-
-    // Get all EndEvents that were run in the Correlation.
-    const endEventInstances = flowNodeInstances.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-
-      const isEndEvent = flowNodeInstance.flowNodeType === BpmnType.endEvent;
-      const isFromProcessModel = flowNodeInstance.processModelId === processModelId;
-
-      // If an onExit token exists, then this FlowNodeInstance was finished.
-      const flowNodeInstanceIsFinished = flowNodeInstance.getTokenByType(ProcessTokenType.onExit) !== undefined;
-
-      // Do not include EndEvent Results from CallActivities or Subprocesses.
-      const isNotFromSubprocess = !flowNodeInstance.parentProcessInstanceId;
-
-      return isEndEvent
-        && isFromProcessModel
-        && flowNodeInstanceIsFinished
-        && isNotFromSubprocess;
-    });
-
-    // Now filter out the EndEvents that the user has no access to.
-    const availableEndEvents = endEventInstances.filter((endEventInstance: FlowNodeInstance): boolean => {
-      return userAccessibleEndEvents
-        .some((accessibleEndEvent: Model.Events.EndEvent): boolean => accessibleEndEvent.id === endEventInstance.flowNodeId);
-    });
-
-    // Now extract all results from the available EndEvents.
-    const results = availableEndEvents.map(this.createCorrelationResultFromEndEventInstance);
-
-    return results;
-  }
-
-  public async getProcessInstancesByIdentity(identity: IIdentity): Promise<Array<DataModels.ProcessInstance>> {
-
-    const suspendedFlowNodeInstances = await this.flowNodeInstanceService.queryActive();
-
-    const flowNodeInstancesOwnedByUser = suspendedFlowNodeInstances.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return this.checkIfIdentityUserIDsMatch(identity, flowNodeInstance.owner);
-    });
-
-    const processInstances = this.processInstanceConverter.convertFlowNodeInstances(flowNodeInstancesOwnedByUser);
-
-    return processInstances;
-  }
-
-  // Events
-  public async getEventsForProcessModel(identity: IIdentity, processModelId: string): Promise<DataModels.Events.EventList> {
-
-    const suspendedFlowNodeInstances = await this.flowNodeInstanceService.querySuspendedByProcessModel(processModelId);
-
-    const suspendedEvents = suspendedFlowNodeInstances.filter(this.isFlowNodeAnEvent);
-
-    const eventList = await this.eventConverter.convertEvents(identity, suspendedEvents);
-
-    return eventList;
-  }
-
-  public async getEventsForCorrelation(identity: IIdentity, correlationId: string): Promise<DataModels.Events.EventList> {
-
-    const suspendedFlowNodeInstances = await this.flowNodeInstanceService.querySuspendedByCorrelation(correlationId);
-
-    const suspendedEvents = suspendedFlowNodeInstances.filter(this.isFlowNodeAnEvent);
-
-    const accessibleEvents = await Promise.filter(suspendedEvents, async (flowNode: FlowNodeInstance): Promise<boolean> => {
-      try {
-        await this.processModelUseCase.getProcessModelById(identity, flowNode.processModelId);
-
-        return true;
-      } catch (error) {
-
-        return false;
-      }
-    });
-
-    const eventList = await this.eventConverter.convertEvents(identity, accessibleEvents);
-
-    return eventList;
-  }
-
-  public async getEventsForProcessModelInCorrelation(
-    identity: IIdentity,
-    processModelId: string,
-    correlationId: string,
-  ): Promise<DataModels.Events.EventList> {
-
-    const suspendedFlowNodeInstances = await this.flowNodeInstanceService.querySuspendedByCorrelation(correlationId);
-
-    const suspendedEvents = suspendedFlowNodeInstances.filter((flowNode: FlowNodeInstance): boolean => {
-
-      const flowNodeIsEvent = this.isFlowNodeAnEvent(flowNode);
-      const flowNodeBelongstoCorrelation = flowNode.processModelId === processModelId;
-
-      return flowNodeIsEvent && flowNodeBelongstoCorrelation;
-    });
-
-    const triggerableEvents = await this.eventConverter.convertEvents(identity, suspendedEvents);
-
-    return triggerableEvents;
-  }
-
-  public async triggerMessageEvent(identity: IIdentity, messageName: string, payload?: DataModels.Events.EventTriggerPayload): Promise<void> {
-
-    await this.iamService.ensureHasClaim(identity, this.canTriggerMessagesClaim);
-
-    const messageEventName = Messages.EventAggregatorSettings.messagePaths.messageEventReached
-      .replace(Messages.EventAggregatorSettings.messageParams.messageReference, messageName);
-
-    this.eventAggregator.publish(messageEventName, payload);
-  }
-
-  public async triggerSignalEvent(identity: IIdentity, signalName: string, payload?: DataModels.Events.EventTriggerPayload): Promise<void> {
-
-    await this.iamService.ensureHasClaim(identity, this.canTriggerSignalsClaim);
-
-    const signalEventName = Messages.EventAggregatorSettings.messagePaths.signalEventReached
-      .replace(Messages.EventAggregatorSettings.messageParams.signalReference, signalName);
-
-    this.eventAggregator.publish(signalEventName, payload);
-  }
-
-  // EmptyActivities
-  public async getEmptyActivitiesForProcessModel(identity: IIdentity, processModelId: string): Promise<DataModels.EmptyActivities.EmptyActivityList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByProcessModel(processModelId);
-
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, suspendedFlowNodes);
-
-    return emptyActivityList;
-  }
-
-  public async getEmptyActivitiesForProcessInstance(
-    identity: IIdentity,
-    processInstanceId: string,
-  ): Promise<DataModels.EmptyActivities.EmptyActivityList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByProcessInstance(processInstanceId);
-
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, suspendedFlowNodes);
-
-    return emptyActivityList;
-  }
-
-  public async getEmptyActivitiesForCorrelation(identity: IIdentity, correlationId: string): Promise<DataModels.EmptyActivities.EmptyActivityList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByCorrelation(correlationId);
-
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, suspendedFlowNodes);
-
-    return emptyActivityList;
-  }
-
-  public async getEmptyActivitiesForProcessModelInCorrelation(
-    identity: IIdentity,
-    processModelId: string,
-    correlationId: string,
-  ): Promise<DataModels.EmptyActivities.EmptyActivityList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByCorrelation(correlationId);
-
-    const suspendedProcessModelFlowNodes = suspendedFlowNodes.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return flowNodeInstance.tokens[0].processModelId === processModelId;
-    });
-
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, suspendedProcessModelFlowNodes);
-
-    return emptyActivityList;
-  }
-
-  public async getWaitingEmptyActivitiesByIdentity(identity: IIdentity): Promise<DataModels.EmptyActivities.EmptyActivityList> {
-
-    const suspendedFlowNodeInstances = await this.flowNodeInstanceService.queryByState(FlowNodeInstanceState.suspended);
-
-    const flowNodeInstancesOwnedByUser = suspendedFlowNodeInstances.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return this.checkIfIdentityUserIDsMatch(identity, flowNodeInstance.owner);
-    });
-
-    const emptyActivityList = await this.emptyActivityConverter.convert(identity, flowNodeInstancesOwnedByUser);
-
-    return emptyActivityList;
-  }
-
-  public async finishEmptyActivity(
-    identity: IIdentity,
-    processInstanceId: string,
-    correlationId: string,
-    emptyActivityInstanceId: string,
-  ): Promise<void> {
-
-    const matchingFlowNodeInstance =
-      await this.getFlowNodeInstanceForCorrelationInProcessInstance(correlationId, processInstanceId, emptyActivityInstanceId);
-
-    const noMatchingInstanceFound = matchingFlowNodeInstance === undefined;
-    if (noMatchingInstanceFound) {
-      const errorMessage =
-        // eslint-disable-next-line max-len
-        `ProcessInstance '${processInstanceId}' in Correlation '${correlationId}' does not have an EmptyActivity with id '${emptyActivityInstanceId}'`;
-      throw new EssentialProjectErrors.NotFoundError(errorMessage);
-    }
-
-    const convertedEmptyActivityList = await this.emptyActivityConverter.convert(identity, [matchingFlowNodeInstance]);
-
-    const matchingEmptyActivity = convertedEmptyActivityList.emptyActivities[0];
-
-    return new Promise<void>((resolve: Function, reject: Function): void => {
-      const routePrameter: {[name: string]: string} = Messages.EventAggregatorSettings.messageParams;
-
-      const emptyActivityFinishedEvent = Messages.EventAggregatorSettings
-        .messagePaths.emptyActivityWithInstanceIdFinished
-        .replace(routePrameter.correlationId, correlationId)
-        .replace(routePrameter.processInstanceId, processInstanceId)
-        .replace(routePrameter.flowNodeInstanceId, emptyActivityInstanceId);
-
-      this.eventAggregator.subscribeOnce(emptyActivityFinishedEvent, (message: InternalActivityFinishedMessage): void => {
-        resolve();
-      });
-
-      this.publishFinishEmptyActivityEvent(identity, matchingEmptyActivity);
-    });
-  }
-
-  // ManualTasks
-  public async getManualTasksForProcessModel(identity: IIdentity, processModelId: string): Promise<DataModels.ManualTasks.ManualTaskList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByProcessModel(processModelId);
-
-    const manualTaskList = await this.manualTaskConverter.convert(identity, suspendedFlowNodes);
-
-    return manualTaskList;
-  }
-
-  public async getManualTasksForProcessInstance(identity: IIdentity, processInstanceId: string): Promise<DataModels.ManualTasks.ManualTaskList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByProcessInstance(processInstanceId);
-
-    const manualTaskList = await this.manualTaskConverter.convert(identity, suspendedFlowNodes);
-
-    return manualTaskList;
-  }
-
-  public async getManualTasksForCorrelation(identity: IIdentity, correlationId: string): Promise<DataModels.ManualTasks.ManualTaskList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByCorrelation(correlationId);
-
-    const manualTaskList = await this.manualTaskConverter.convert(identity, suspendedFlowNodes);
-
-    return manualTaskList;
-  }
-
-  public async getManualTasksForProcessModelInCorrelation(
-    identity: IIdentity,
-    processModelId: string,
-    correlationId: string,
-  ): Promise<DataModels.ManualTasks.ManualTaskList> {
-
-    const flowNodeInstances = await this.flowNodeInstanceService.queryActiveByCorrelationAndProcessModel(correlationId, processModelId);
-
-    const suspendedFlowNodeInstances = flowNodeInstances.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return flowNodeInstance.state === FlowNodeInstanceState.suspended;
-    });
-
-    const manualTaskList = await this.manualTaskConverter.convert(identity, suspendedFlowNodeInstances);
-
-    return manualTaskList;
-  }
-
-  public async getWaitingManualTasksByIdentity(identity: IIdentity): Promise<DataModels.ManualTasks.ManualTaskList> {
-
-    const suspendedFlowNodeInstances = await this.flowNodeInstanceService.queryByState(FlowNodeInstanceState.suspended);
-
-    const flowNodeInstancesOwnedByUser = suspendedFlowNodeInstances.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return this.checkIfIdentityUserIDsMatch(identity, flowNodeInstance.owner);
-    });
-
-    const manualTaskList = await this.manualTaskConverter.convert(identity, flowNodeInstancesOwnedByUser);
-
-    return manualTaskList;
-  }
-
-  public async finishManualTask(
-    identity: IIdentity,
-    processInstanceId: string,
-    correlationId: string,
-    manualTaskInstanceId: string,
-  ): Promise<void> {
-
-    const matchingFlowNodeInstance =
-      await this.getFlowNodeInstanceForCorrelationInProcessInstance(correlationId, processInstanceId, manualTaskInstanceId);
-
-    const noMatchingInstanceFound = matchingFlowNodeInstance === undefined;
-    if (noMatchingInstanceFound) {
-      const errorMessage =
-        `ProcessInstance '${processInstanceId}' in Correlation '${correlationId}' does not have a ManualTask with id '${manualTaskInstanceId}'`;
-      throw new EssentialProjectErrors.NotFoundError(errorMessage);
-    }
-
-    const convertedUserTaskList = await this.manualTaskConverter.convert(identity, [matchingFlowNodeInstance]);
-
-    const matchingManualTask = convertedUserTaskList.manualTasks[0];
-
-    return new Promise<void>((resolve: Function, reject: Function): void => {
-      const routePrameter: {[name: string]: string} = Messages.EventAggregatorSettings.messageParams;
-
-      const manualTaskFinishedEvent = Messages.EventAggregatorSettings
-        .messagePaths.manualTaskWithInstanceIdFinished
-        .replace(routePrameter.correlationId, correlationId)
-        .replace(routePrameter.processInstanceId, processInstanceId)
-        .replace(routePrameter.flowNodeInstanceId, manualTaskInstanceId);
-
-      this.eventAggregator.subscribeOnce(manualTaskFinishedEvent, (message: InternalActivityFinishedMessage): void => {
-        resolve();
-      });
-
-      this.publishFinishManualTaskEvent(identity, matchingManualTask);
-    });
-  }
-
-  // UserTasks
-  public async getUserTasksForProcessModel(identity: IIdentity, processModelId: string): Promise<DataModels.UserTasks.UserTaskList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByProcessModel(processModelId);
-
-    const userTaskList = await this.userTaskConverter.convertUserTasks(identity, suspendedFlowNodes);
-
-    return userTaskList;
-  }
-
-  public async getUserTasksForProcessInstance(identity: IIdentity, processInstanceId: string): Promise<DataModels.UserTasks.UserTaskList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByProcessInstance(processInstanceId);
-
-    const userTaskList = await this.userTaskConverter.convertUserTasks(identity, suspendedFlowNodes);
-
-    return userTaskList;
-  }
-
-  public async getUserTasksForCorrelation(identity: IIdentity, correlationId: string): Promise<DataModels.UserTasks.UserTaskList> {
-
-    const suspendedFlowNodes = await this.flowNodeInstanceService.querySuspendedByCorrelation(correlationId);
-
-    const userTaskList = await this.userTaskConverter.convertUserTasks(identity, suspendedFlowNodes);
-
-    return userTaskList;
-  }
-
-  public async getUserTasksForProcessModelInCorrelation(
-    identity: IIdentity,
-    processModelId: string,
-    correlationId: string,
-  ): Promise<DataModels.UserTasks.UserTaskList> {
-
-    const flowNodeInstances = await this.flowNodeInstanceService.queryActiveByCorrelationAndProcessModel(correlationId, processModelId);
-
-    const suspendedFlowNodeInstances = flowNodeInstances.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return flowNodeInstance.state === FlowNodeInstanceState.suspended;
-    });
-
-    const noSuspendedFlowNodesFound = !suspendedFlowNodeInstances || suspendedFlowNodeInstances.length === 0;
-    if (noSuspendedFlowNodesFound) {
-      return <DataModels.UserTasks.UserTaskList> {
-        userTasks: [],
-      };
-    }
-
-    const userTaskList = await this.userTaskConverter.convertUserTasks(identity, suspendedFlowNodeInstances);
-
-    return userTaskList;
-  }
-
-  public async getWaitingUserTasksByIdentity(identity: IIdentity): Promise<DataModels.UserTasks.UserTaskList> {
-
-    const suspendedFlowNodeInstances = await this.flowNodeInstanceService.queryByState(FlowNodeInstanceState.suspended);
-
-    const flowNodeInstancesOwnedByUser = suspendedFlowNodeInstances.filter((flowNodeInstance: FlowNodeInstance): boolean => {
-      return this.checkIfIdentityUserIDsMatch(identity, flowNodeInstance.owner);
-    });
-
-    const userTaskList = await this.userTaskConverter.convertUserTasks(identity, flowNodeInstancesOwnedByUser);
-
-    return userTaskList;
-  }
-
-  public async finishUserTask(
-    identity: IIdentity,
-    processInstanceId: string,
-    correlationId: string,
-    userTaskInstanceId: string,
-    userTaskResult?: DataModels.UserTasks.UserTaskResult,
-  ): Promise<void> {
-
-    const resultForProcessEngine = this.createUserTaskResultForProcessEngine(userTaskResult);
-
-    const matchingFlowNodeInstance =
-      await this.getFlowNodeInstanceForCorrelationInProcessInstance(correlationId, processInstanceId, userTaskInstanceId);
-
-    const noMatchingInstanceFound = matchingFlowNodeInstance === undefined;
-    if (noMatchingInstanceFound) {
-      const errorMessage =
-        `ProcessInstance '${processInstanceId}' in Correlation '${correlationId}' does not have a UserTask with id '${userTaskInstanceId}'`;
-      throw new EssentialProjectErrors.NotFoundError(errorMessage);
-    }
-
-    const convertedUserTaskList = await this.userTaskConverter.convertUserTasks(identity, [matchingFlowNodeInstance]);
-
-    const matchingUserTask = convertedUserTaskList.userTasks[0];
-
-    return new Promise<void>((resolve: Function, reject: Function): void => {
-
-      const userTaskFinishedEvent = Messages.EventAggregatorSettings.messagePaths.userTaskWithInstanceIdFinished
-        .replace(Messages.EventAggregatorSettings.messageParams.correlationId, correlationId)
-        .replace(Messages.EventAggregatorSettings.messageParams.processInstanceId, processInstanceId)
-        .replace(Messages.EventAggregatorSettings.messageParams.flowNodeInstanceId, userTaskInstanceId);
-
-      this.eventAggregator.subscribeOnce(userTaskFinishedEvent, (message: InternalUserTaskFinishedMessage): void => {
-        resolve();
-      });
-
-      this.publishFinishUserTaskEvent(identity, matchingUserTask, resultForProcessEngine);
-    });
-  }
-
-  private async getFlowNodeInstanceForCorrelationInProcessInstance(
-    correlationId: string,
-    processInstanceId: string,
-    instanceId: string,
-  ): Promise<FlowNodeInstance> {
-
-    const suspendedFlowNodeInstances = await this.flowNodeInstanceService.querySuspendedByProcessInstance(processInstanceId);
-
-    const matchingInstance = suspendedFlowNodeInstances.find((instance: FlowNodeInstance): boolean => {
-      return instance.id === instanceId &&
-             instance.correlationId === correlationId;
-    });
-
-    return matchingInstance;
-  }
-
-  private createCorrelationResultFromEndEventInstance(endEventInstance: FlowNodeInstance): DataModels.CorrelationResult {
-
-    const exitToken = endEventInstance.tokens.find((token: ProcessToken): boolean => {
-      return token.type === ProcessTokenType.onExit;
-    });
-
-    const correlationResult: DataModels.CorrelationResult = {
-      correlationId: exitToken.correlationId,
-      endEventId: endEventInstance.flowNodeId,
-      tokenPayload: exitToken.payload,
-    };
-
-    return correlationResult;
-  }
-
-  private isFlowNodeAnEvent(flowNodeInstance: FlowNodeInstance): boolean {
-    return flowNodeInstance.eventType !== undefined;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  private createUserTaskResultForProcessEngine(finishedTask: DataModels.UserTasks.UserTaskResult): any {
-
-    const noResultsProvided = !finishedTask || !finishedTask.formFields;
-
-    if (noResultsProvided) {
-      return {};
-    }
-
-    const formFieldResultIsNotAnObject = typeof finishedTask !== 'object'
-      || typeof finishedTask.formFields !== 'object'
-      || Array.isArray(finishedTask.formFields);
-
-    if (formFieldResultIsNotAnObject) {
-      throw new EssentialProjectErrors.BadRequestError('The UserTask\'s FormFields are not an object.');
-    }
-
-    return finishedTask.formFields;
-  }
-
-  private checkIfIdentityUserIDsMatch(identityA: IIdentity, identityB: IIdentity): boolean {
-    return identityA.userId === identityB.userId;
-  }
-
-  private publishFinishEmptyActivityEvent(
-    identity: IIdentity,
-    emptyActivityInstance: DataModels.EmptyActivities.EmptyActivity,
-  ): void {
-
-    const finishEmptyActivityMessage = new InternalFinishEmptyActivityMessage(
-      emptyActivityInstance.correlationId,
-      emptyActivityInstance.processModelId,
-      emptyActivityInstance.processInstanceId,
-      emptyActivityInstance.id,
-      emptyActivityInstance.flowNodeInstanceId,
-      identity,
-      emptyActivityInstance.tokenPayload,
-    );
-
-    const finishEmptyActivityEvent = Messages.EventAggregatorSettings.messagePaths.finishEmptyActivity
-      .replace(Messages.EventAggregatorSettings.messageParams.correlationId, emptyActivityInstance.correlationId)
-      .replace(Messages.EventAggregatorSettings.messageParams.processInstanceId, emptyActivityInstance.processInstanceId)
-      .replace(Messages.EventAggregatorSettings.messageParams.flowNodeInstanceId, emptyActivityInstance.flowNodeInstanceId);
-
-    this.eventAggregator.publish(finishEmptyActivityEvent, finishEmptyActivityMessage);
-  }
-
-  private publishFinishManualTaskEvent(
-    identity: IIdentity,
-    manualTaskInstance: DataModels.ManualTasks.ManualTask,
-  ): void {
-
-    // ManualTasks do not produce results.
-    const emptyPayload = {};
-    const finishManualTaskMessage = new InternalFinishManualTaskMessage(
-      manualTaskInstance.correlationId,
-      manualTaskInstance.processModelId,
-      manualTaskInstance.processInstanceId,
-      manualTaskInstance.id,
-      manualTaskInstance.flowNodeInstanceId,
-      identity,
-      emptyPayload,
-    );
-
-    const finishManualTaskEvent = Messages.EventAggregatorSettings.messagePaths.finishManualTask
-      .replace(Messages.EventAggregatorSettings.messageParams.correlationId, manualTaskInstance.correlationId)
-      .replace(Messages.EventAggregatorSettings.messageParams.processInstanceId, manualTaskInstance.processInstanceId)
-      .replace(Messages.EventAggregatorSettings.messageParams.flowNodeInstanceId, manualTaskInstance.flowNodeInstanceId);
-
-    this.eventAggregator.publish(finishManualTaskEvent, finishManualTaskMessage);
-  }
-
-  private publishFinishUserTaskEvent(
-    identity: IIdentity,
-    userTaskInstance: DataModels.UserTasks.UserTask,
-    userTaskResult: DataModels.UserTasks.UserTaskResult,
-  ): void {
-
-    const finishUserTaskMessage = new InternalFinishUserTaskMessage(
-      userTaskResult,
-      userTaskInstance.correlationId,
-      userTaskInstance.processModelId,
-      userTaskInstance.processInstanceId,
-      userTaskInstance.id,
-      userTaskInstance.flowNodeInstanceId,
-      identity,
-      userTaskInstance.tokenPayload,
-    );
-
-    const finishUserTaskEvent = Messages.EventAggregatorSettings.messagePaths.finishUserTask
-      .replace(Messages.EventAggregatorSettings.messageParams.correlationId, userTaskInstance.correlationId)
-      .replace(Messages.EventAggregatorSettings.messageParams.processInstanceId, userTaskInstance.processInstanceId)
-      .replace(Messages.EventAggregatorSettings.messageParams.flowNodeInstanceId, userTaskInstance.flowNodeInstanceId);
-
-    this.eventAggregator.publish(finishUserTaskEvent, finishUserTaskMessage);
-  }
-
-}
