+++ src/index.ts
@@ -1,4 +1,9 @@
-export * from './consumer_api_service';
-
 export * from './adapters/index';
 export * from './converters/index';
+
+export * from './empty_activity_service';
+export * from './event_service';
+export * from './manual_task_service';
+export * from './notiifcation_service';
+export * from './process_model_service';
+export * from './user_task_service';
